pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-p2p-snapshot"
version = "0.1.0"
description = "Decentralized Nano ledger snapshot distribution via BitTorrent BEP 46"
requires-python = ">=3.12"
dependencies = [
    "pynacl>=1.5",
    "bencodepy>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
nano-producer = "producer.cli:main"
nano-mirror = "mirror.watcher:main"
nano-push-status = "producer.push_status:main"

[tool.setuptools.packages.find]
include = ["shared*", "mirror*", "producer*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
